requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
docker==7.1.0
executing==2.1.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
icecream==2.1.3
idna==3.10
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
pytest-alembic==0.11.1
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
